LayoutTests/inspector/console/console-uncaught-exception-expected.txt

CONSOLE MESSAGE: line 18: Error: Exception in inline script.
CONSOLE MESSAGE: line 11: Error: Exception in 'load' event listener.
CONSOLE MESSAGE: line 6: Error: Exception in setTimeout callback.
Tests that uncaught exceptions are logged into console.Bug 47250.


Error: Exception in inline script. resources/uncaught-in-iframe.html:18a resources/uncaught-in-iframe.html:18b resources/uncaught-in-iframe.html:23
Error: Exception in 'load' event listener. resources/uncaught-in-iframe.html:11f resources/uncaught-in-iframe.html:11(anonymous function) resources/uncaught-in-iframe.html:13
Error: Exception in setTimeout callback. uncaught-in-iframe.html:6bar uncaught-in-iframe.html:6(anonymous function) uncaught-in-iframe.html:8
